git-svn-id: http://svn.boost.org/svn/boost/branches/release@63664 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/graph/distributed/rmat_graph_generator.hpp
@@ -60,7 +60,7 @@ namespace boost {
           if (permute_vertices) 
               generate_permutation_vector(gen, vertexPermutation, n);
 
-          int SCALE = int(floor(log(n)/log(2.)));
+          int SCALE = int(floor(log(double(n))/log(2.)));
           boost::uniform_01<RandomGenerator> prob(gen);
       
           std::map<value_type, bool> edge_map;
